src/gem-3_1: Added file
Well, this is the `war3map.j` for the last official version of the map:
`[BK's] Gem Tower Defense v3.1`.  The script was indented and strings were
inlined.  It should be mentioned that a number of the quest messages were too
long, and needed concatenation to keep string sizes under the 1013 save/load
limit.