Issue #2: Fix downgraded Opal aura level
A downgraded Opal always had a Level 1 aura, no matter the original quality.
Apparently the `-debug` command addresses this issue, and is the correct fix.
However, it is not applied in the proper location, and so in games where the
command is not used the Opals continue to have the wrong aura.

This deals with the issue immediately after an Opal is downgraded.