Issue #30: Support Grim Extension Pack
First off, the extension pack itself is added into the repository.  Like with
`pjass`, this will make builing the map easier for other people.

Second, the build script and configuration table have been updated to handle
merging objects into a specified map file.  This is nice as it will allow
tracking of changes to object data.  Also, it allows the ability to remove
objects from the map file, and use this process to define them.  Eventually
all objects could be handled in such a fashion.