Issue #26: Added unit indexing
Opted for a system that pretty much is self contained, as opposed to hooking
various functions such as `CreateUnit ()` and `RemoveUnit ()`.  Additionally,
makes use of GrimEx to add its needed ability to the map.

Note that this addition actually breaks the map, as the map still uses the
functions `GetUnitUserData ()`, and `SetUnitUserData ()`.  As a result, a
system that stimulates unit user data is added, and any calls have been
updated accordingly.