Issue #4: Kills not registering properly
The issue revolves around triggers that fire with `EVENT_PLAYER_UNIT_DEATH`.
Conditions for the bug to occur are rare, and require combining when a part
of the combination lands a killing blow.  However, the killing unit does not
exist because it was replaced, so the kill can't be properly registered.

The fix involved storing the owner of each creep that is spawned, and using
that later to assign the kill to the proper player.  The solution also deals
with another rare bug that involves a player killing the creep of another
player.

Do note that this does not address a similar issue that causes kills to not
increment their respective structure's kill total.  That is not a game
breaking bug, and would require a more complex solution.