src/settings: Reworked timer handling
Rather than attempting to handle an anonymous timer by calling
`GetExpiredTimer ()`, a global referencing the timer is used.  This is because
attempting to get an expired timer when no time was used will crash the game.
So this is a bit safer, and does not require the use of a wrapper function.