Issue #46: Empty board rows
This resolves the issue of missing players occupying a row in the board.  Now,
only players that begin the game will occupy a row.  If a player leaves,
information about them will remain, but not be updated.

Also, the board for `Classic` and `Race` mode has become exactly the same, to
simplify things.  Given that `Classic` mode represent a miniscule fraction of
the total games, I doubt anyone will notice.  Also, if `Classic` disappears
entirely, nothing has to be done in regards to the board.