Issue #57: Kill count off after Rerunner
This bug only appears in `Race` mode.  It seems to be a very old bug that was
never encountered until after the fix applied by #4.  In fact, before #4
everything ran as expected.  The reason why is because the old behavior caused
the trigger handling kills in `Race` mode to ignore the Rerunner as a check on
the kiling unit returned a computer player, given ownership had been
transfered by that point for gem placement.  After the fix, however, the check
began to pass, and thus the next round began with one unit already dead, and
all subsequent ones started and ended too soon.

Note that `Classic` mode was not tested before or after this fix.  However, it
looks as though everything should be fine.  Then again, that's what I said
before this bug showed up.