Issue #24: Remove `Classic` mode
Simply too slow.  I have no desire to maintain this mode, and so it is going.
I thought I would initially keep it in, but I find myself having to fix bugs
twice, due to the code duplication involved in supporting both modes.  Now, I
could merge the code together, but that is a task in and of itself.  Not to
mention I simply do not enjoy `Classic` mode, and I'm pretty sure no one even
plays it anymore.  Overall this cleans up the code base a bit, and will make
managing things simpler.