Issue #8: Fix combining
The issue here is that when combining a piece is randomly selected to fufill
the recipe.  When done immediately it is possible that a newly selected piece
will be used to fulfill an existing combination before said piece can be
properly registered.  Reorganizing the code to not be stuck behind waits fixes
this partially.

Another potential case exists when gathering multiple full sets of a recipe,
then trying to combine.  Fulfilling one combination could use the 'combining'
piece of the other, thus making it impossible to finish that set.  Adding a
check to ensure that pieces containing the `Combine` ability are not used
resolves this issue.