Extra chancing finally fixed
Issue: #60

The fix for #60 neglected to include a line to reset the `boolean` flag used
by the players.  As a result extra chancing was guaranteed to break the moment
they used it for anything besides slates.

On top of that, the way the GUI generated code worked allowed each check to
occur, and as such additional blocks were being executed and giving players
extra gold.