Game timer now starts when game is initialized
Issue: #66

Previously it started immediately after the map loaded.  This was not much of
a problem when the game was started using an HCL command.  However, when a
player selected and confirmed the modes a delay could occur, thus making the
game clock out of sync with the actual amount of time a player spent building
their maze.