Update map name and descriptions
Issue: #1, #64

Changing the map name to `Gem TD+` took a bit of thinking.  This remains
simple and indicates that the map is essentially starting over and being
actively maintained again.  Also, there is a bit of separation between the
original line and this, so perhaps people will be less opposed to changes.

Updating the loading screens is being done by hand.  It would be nice to
automate the process, but for now this is fine.