Prevent the quests tab from flashing
Issue: #71

The culprit was `QuestMessageBJ ()`.  The solution was to simply copy the
function and remove the behavior that causes the quests tab to flash.

I considered preventing the quest sounds from playing, but until I hear enough
people complaining about them I will do nothing.  I do not play the map with
sounds on, so I have no clue if they are annoying or not.  Or even if they're
helpful.