Rework game start
Issues: #79, #84, #95

Given that when bot hosted via HCL the text displayed by the map was not
easily visible, that text is now removed in favor of timer dialogs.
Additionally, in light of possible camera shifting at game start, locking of
user control is more consistent.  On top of that the player's camera is now
locked to their miner until a forced game wait elapses.

This forced wait was deemed necessary given that without it HCL games
essentially started immediately without any real warning.  Also, it gives
players with slower computers a little extra time for things to settle.