Return player control at start of game
Issues: #84, #95, #100

After adding the pause in #95 it was noticed that being unable to do anything
felt a bit jarring.  The previous pause during `Gem TD 3.1` settings selection
was not as much of an issue, as it was typically very short.  And even though
use of `EnableUserControler ()` was carried over from the previous versions
and remained in an attempt to address camera shifting in #84, it was apparent
that this was all futile in the end.  There is a delay in the dialog button
events that is unavoiable, and as such it is more or less impossible to
prevent potential camera issues.

In light of the above, player control has been returned during both game
settings selection and the game start timer.  Instead, pausing the Miner for
each player has become the favored solution.  This prevents the unit from
being moved; however, other actions, which previously were disabled, can
occur: the ability to type; accessing game tabs and menus; use of the `-zoom`
command; setting control groups, etc.

Of course, these are for the player not selecting the game settings.  If the
game start timer did not exist, this player would be at a slight disadvantage,
having not had the opportunity to perform the above mentioned actions.  As
such, the game start timer will remain, if for no other reason than it gives
the selecting player time to be on equal footing with the other players.