Cleanup settings selection window
For starters, a rename of dialog to window internally.  Beyond this, a big
change was shifting the creation of objects from being inside the globals
block to being only when the system is initialized via
`Settings_Window__Initialize ()`.  In other words, when the settings window is
now needed (i.e. HCL hosted) then objects related to it will not be created.

Additionally, the menu for the window will have its labels cached, thus making
refreshing the menu faster.  On top of that, a new option has been added to
the menu: the ability to pass one's turn, thus eliminating the need to have to
wait the full duration of their turn to accomplish the same feat.