Fix/update the character width system
Issues: #99, #110

For starters, the character width system can now handle unknown characters.
When encountered the returned value is equal to the widest character.  It is
better to err on the side of caution.  And when an empty string or `null` is
encountered, zero is returned.

Also, the calculated values for the characters were apparently wrong.  They
have been updated accordingly.