Update build script to use an optimizer
Firstly, the base file is updated to match that in the configuration.  It is
nothing more than a symbolic link that points to `[BK's] Gem TD v3.1.w3x`,
which is a shell of the original file, lacking a `war3map.j`.

Next, Vexorian's Optimizer 5.0 is added to the repository.  This is used via
the build script to optimize the map file present in the `output_directory` of
the configuration file (after using the `data` command).  The destination will
be suffixed with `-optimized`, and will be created in the same directory.