Document the build process
Issues: #20, #117

This finally documents how to build the map.  Sadly, much is left unautomated
still.  Ideally a single build script would be used to create an optimized
version of the map, but not yet.  For now we still have the file mentioning
changes that must be done by hand, which has been extensively updated.

Also, documentation regarding how to verify the contents `war3map.j` is added,
for those individuals who doubt the intentions of this project.