Updated issue labels documentation
Removed priority labels, as they really did not serve much purpose in my work
flow.  Also, I do not want people to think their issue is more or less
importmant, because use of a priority label was purely to help me organize.

Also, changed `Area` to `Scope` and `Type` to `Category` in an attempt to
ensure that the first category alphabetically relates to the type of label.
This will ensure that `Category: Disccusion` is much closer to the top,
helping to facilitate discourse.