Resolve leaks related to special effects
Issue: #136

A hunch tells me that this is the biggest contributor to the performance
issues experienced towards the end of long games.  Both special effects and
locations were leaked throughout the game, albeit I am sure the special
effects played the larger part.

A worry of mine was that there would be more special effects that could not be
destroyed immediately via `DestroyEffect ()`.  However, only two were found to
be subject to this constraint:

- The initial indication of finding a combination (a whitish glow at the
  location);
- And the constant orange pattern underneath a structure that can be used to
  fulfill a recipe.

Given the non-critical nature of these, and the fact that the second is
already on a timer that creates a new effect every `3.5` seconds,
`TriggerSleepAction ()` was deemed sufficient.  The function is not lag or
pause resistant; however, it is good enough for this application, and very
simple to implement.

Do note the use of `ExecuteFunc ()` within a `ForGroup ()`.  This allows the
call to `TriggerSleepAction ()` to succeed, whereas otherwise the call would
have halted.