Adjust introduction and settings selection wording
Issue: #74, #138

In both cases this serves as clarification to the players.  In regards to the
introduction, the fact that they can actually type turn the game start timer.
This may not be obviously apparent to many, given Gem's previous nature of
disabling the user interface.

In regards to settings selection, it is more for the person selecting the
settings than anything.  This is because it is impossible for the other
players to tell them what to do, given how player chat messages work in
conjunction with an open dialog window.