Modify refund handling for canceling construction
Issue: #140

Currently the code tries to 'idiot proof' building cancellation, specifically
for gem placement.  However, in dealing with #107 another constant was seen:
`Cancel Building Factor - Building Construction`.  The default value for this
is `0.75`; this fix changes that to `1.00`.  As a result, the existing code is
rendered redundant.  However, every other building the player can construct
must have a lumber cost of `0`, or they could get a refund when one is not
expected.

All towers and structures that the player can 'build' now have their lumber
cost set to `0`.  Although, this adjustment may not have been entirely
necessary.