Valid Warcraft 3 install is a build requirement
Issue: #142

A valid Warcraft 3 installation has always been a requirement, even for
version `1.0`, regardless of whether or not it was mentioned.  This was due to
the use of GrimEx.  This requirement was simply overlooked due to the fact
that if someone was planning to play the map (e.g. myself), then it is highly
likely they they have a working installation.