Fix issues preventing save/load compatibility
Issue: #155

After researching this bug I have come to the conclusions listed below. First
off, this is not a definitive list of what breaks save/load functionality.
This is a list of what broke save/load functionality in Gem TD+.

Also included in the list are things that reportedly break save/load
functionality, but were tested and found to not be true. Mind you, the
validity of these claims could depend on the patch version applied to Warcraft
 3. Gem TD+ is only concerned with the latest patch version (currently
`1.27a`).

- A bit of testing confirms that using index `8191` within an array will cause
  a crash when loading a saved game. The original Gem TD never used indices
  this high, so this behavior was previously never encountered.
- A source said that periodic functions do not work properly with save/load.
  Supposedly, they become 'one-shot' after loading. Perhaps this was the case
  in a previous patch version; however, my testing shows they work just fine.
- Another source said that to provide save/load compatibility one cannot
  exceed a string with a length of `1013` characters. My own testing has
  shown that the upper limit on string length is `1023` characters, the same
  limit Warcraft 3 has for string length in general. To check string length I
  used the function `StringLength ()`, and printed the results. My test
  included color codes and line breaks (`\n`). I also tested with a string
  that had no color code or line breaks, in case those are not counted towards
  the limit (I am almost certain they are).

There were two areas in Gem TD+ that were causing save/load functionality to
break:
- The unit indexing system; and,
- The quest/information system.

In regards to unit indexing, a simple adjustment to not use index `8191` with
a particular array solves the problem. The system has been updated to reflect
this information.

With the quest/information system, the issue was the use of strings that
exceeded the `1023` character limit. Mind you, directly declaring a string in
excess of `1023` characters in JASS will cause the map to crash. However, you
can get around this by concatenating valid strings together, then passing
those to the native functions. This is what Gem TD+ was doing.

The problem arises when using save/load. I imagine that the full version of a
string is saved, not the smaller parts that may be initially concatenated.
Thus, when loading, the game encounters a single string that exceeds the
`1023` character limit, and subsequently crashes.

Curiously, the original versions of Gem TD also had long quest messages that
were certainly over the `1023` character limit. However, they made use of the
`war3map.wts` file, rather than essentially inlining strings directly into the
`war3map.j`. I have no plans to test this, but I would presume that save/load
does not touch strings within the `war3map.wts`, and simply includes the file.
Thus, long strings (those exceeding `1023` characters) within it do not break
save/load functionality, as they do not seem to cause the game to crash.

The question/information system has been rewritten, and will now display an
error if a quest description is too long. This error is primarily intended for
development purposes, as its mere presence is an indication of save/load
incompatibility.