Split `Unit Event` from `Unit Indexer`
This change is not entirely needed, but provides slightly easier to maintain
code. Furthermore, an attempt is made to improve comments through the files
involved.

A curious thing to note is that the stack implementation for `Unit Indexer`
was bugged. Essentially, if the first unit popped was then pushed back, an
erroneous 'overflow' error would have been encountered. This bug must have
been introduced when I converted from a queue to a stack during early
development.

However, the bug does not affect previous versions of Gem TD+ given that many
units are indexed at the start of the game before a single unit leaves the
map. And, for all units to be deindexed, every player must leave, thus causing
the game to end before such an even would ever occur.