Fix problem handling multiple 'Combine' orders
Issue: #160

The main issue here is that the existing code seems unable to handle multiple
'Combine' orders properly. Note that this only happens when attempting to use
'Combine' when there are multiple targets of the same type and quality alive
(e.g. two Chipped Rubies). Multiple orders are handled improperly, in no small
part to the use of `TriggerSleepAction ()`, and ultimately the player is left
with five rocks and no combined gem.

This bug does not occur when there is only one target alive, nor for 'Keep',
nor for 'Special', nor for 'Downgrade'. All of those function properly, even
when given multiple concurrent orders.

The solution was fairly simple: remove the use of `TriggerSleepAction ()`.
This ensures that the current thread does not halt, which was the issue behind
this particular bug. Additionally, a few unnecessary instances of
`TriggerSleepAction ()` were removed throughout the code base.