Address mouse selection bug
Issue: #159

First off, this is not an outright fix. This merely adds the ability to fix
the bug via the `-debug` command by using the `EnableSelect ()` function.

As for the bug itself, it is a old Warcraft 3 bug with an unknown cause. It
happens in other maps from time to time, but no one has pinpointed the cause.
It is possible that something in the code causes it to occur, but I cannot be
sure. Eventually this may be properly resolved, but until then players at
least can use `-debug`.