Select resulting gem after numerous actions
Issue: #148

The issue primarily reffered to the fact that the result of combining a recipe
did not select the result. However, it was discovered that other actions (i.e.
creating slates, combining recipes, double combine, quad combine, one-hits,
and downgrading) did not select their result the same way keeping a placed gem
did. This commit fixes that behavior, and causes all of the above listed
actions to select their result immediately after the ability is used.