Exclude moved slate from slate stack checks
Issue: #166

Previous, a slate being moved was included in checks to determine if slates
were stacking. This prevented a slate from being moved close to itself,
because the code believed the slate was stacking with itself.

This addresses that behavior, and excludes the slate being moved from the
checks. This allows that slate to be moved close to itself, so long as it is
not stacking with other nearby slates.