Address bugs created by being overzealous
Issues: #170

This partially reverts commit 1fabedaf22efe765db432fa09b839e4b8fa0ac66.

Typically, I have been trying to approach the code base by only making the
bare minimum number of changes to reach the desired goal. But given how many
issues are related to `TriggerSleepAction ()`, you could say I got a little
trigger happy with the above commit in addressing #160.

A side effect of removing the unnessary changes is that the following two
commits are no longer needed (#165):

This reverts commit 88d92c5cab39c7188336b58ce00a60fcf358d270.
This reverts commit 66c95f2732d4cae882f61cb673677b1dcce62e40.

This has shown me how vigilant I must be when addressing future issues, given
the fragility of the existing code base. But what is really needed is a
rewrite of almost the entire map. Until then, steps will be taken to be as
delicate as possible.