Address a crash related to slates
Issue: #169

This reverts a single line in 24f4cf613614995f564168bb0c44de6461d4b5ae.

Okay, so this crash is a direct result of fixing #75. The removal of a
`TriggerSleepAction ()` before the removal of `'A02J'` causes the game to
crash when a player attempts to move and then combine quickly. This fixes this
problem.