Update race completion text
Issue: #175

Officially you complete race mode whether you killed the Damage Test or not.
However, it only displayed that information if you stayed the entire duration
of your Damage Test.

This change makes it so everyone is informed that a player has complete race
mode as the Damage Test is spawned.