Add a system to handle stunning units
Issue: #174

It was decided to keep this system fairly simply. It tracks whether or not a
unit is stunned, and attaches a timer that expires after the stun duration is
up. Anything else that is needed can be implemented externally to this system.

As far as the objects, numerous systems use Firebolt as their base ability.
Also, multiple systems have various tweaks and fixes to ensure that Firebolt
performs as expected: Wurst; Netsharus; Status; etc.