Fix handling of multiple 'Create Slate' orders
Issues: #191, #196

This issue only happens when both the following conditions are met:

- There are multiple possible choices that have the 'Create Slate' ability.
  For example, multiple Amethyst that can create an Air Slate.
- These gems are selected at the same time, and the 'Create Slate' ability is
  spammed rapidly.

The result is that multiple of the same round are started concurrently,
causing extra units to be spawned. Killing said units, due to the current
design of Gem, can cause multiple rounds to be skipped.

The cause of the problem is, yet again, `TriggerSleepAction ()`. The fix for
this exact issue was unknowingly committed when dealing with #160. However,
the initial commit was reverted, and the changes were never applied. Realize
that I was simply removing a bunch of `TriggerSleepAction ()` calls that ran
into, for no other reason than to remove them. I didn't realize this bug even
existed then.

Anyways, this commit removes them, and nothing else.