Make use of the new unit stun and disarm systems
Issues: #154, #156

All of the sources of stun in the map (i.e. Jade, Dark Emerald, Hold Slate,
and Ancient Slate) have been rewritten to make use of the unit stun and disarm
systems. This should address numerous issues that existed in the map.

The existing setup used in the map to deal with stunning units and preventing
attacks used a combination of `TriggerSleepAction ()` and `PauseUnit ()`. The
waits are susceptible to lag and pause, and thus have a wide variability in
their accuracy. The unit pause itself presents an issue in how it interacts
with other game functionality. In particular, certain area of effect damages
did not apply to paused units at all.

Now, the existing setup was fine in how it tracked what was stunned and what
was not. The use of `GetUnitAcquireRange ()` and `SetUnitAcquireRange ()`
looked completely fine. However, in light of the functionality added with the
unit stun system, this behavior has been rendered obsolete.