Adds a spawn system
Essentially, this allows allocation of spawn objects that can be used to
create units according to a specific configuration. The various settings
include the owner, unit type, unit count per wave, number of waves, location,
unit facing, delay before first wave, and the period between waves.

See the documentation in the file for the spawn objects for more details.