Add a spawn group system
This is based upon `Spawn`, the spawn object system. Essentially, this
presents a way to manage a group of spawn objects. It functions in a similar
fashion, but does make an assumption that certain information is shared and
synchronized across the entire group. Also, existing spawn objects cannot be
added to this system. Despite these limitations, it should be quite useful
when a single spawn object is not sufficient.