Dialog button use can disable user control
Issue: #159

The issue is passing multiple dialog button events at the same time, typically
due to spamming the button during lag. The result is that user's control being
disabled.

This is an ancient Warcraft 3 bug. Any map using dialog windows potentially
suffers from this issue. To be fair, however, it is quite rare. The solution
is simple: use of the function `EnableUserControl ()` upon handling a dialog
button event.