Resolve desync issue when using 'Keep'
Issue: #215

This was most likely added in #148. I believe not enough thought was put into
proper handling around the `TriggerSleepAction ()` calls, given the previous
experiences when dealing with matters that led up to #170.

But this must be a very rare desync. I could not duplicate the behavior in
testing. However, a bit of a code rewrite was called for regardless.