Fix taunt for Hold and Ancient slates
Issue: #213

This was fairly simple to figure out. Using `IssueTargetOrder ()` (and perhaps
other group ordering functions) seems to deal with a maximum of twelve units
at a time. However, the taunts for the slates need to handle more than those.

The fix is to simply order the units to attack during group enumeration,
instaed of adding them to the group.