Decrease spawn frequency
Due to the use of `TriggerSleepAction ()` in the old setup, the wait of `1.50`
turns out to not be enough when using a more accurate timer. The previous
behavior can never be exactly replicated, without introducing a variability to
the spawn frequency of `+/-0.15` or so (or `+/- 1.50` under extreme lag.
However, I'm happy with a nice steady spawn right now.