Slates get damaged during ReRunner
Issue: #222

Okay, so this is an old bug. Spent a fair bit of time digging in the Object
Editor, which ended up being unncessary. However, the use of Spiked
Carapace by the ReRunner for the armor reduction is what causes this behavior.

Essentially, what is happening is that the function `UnitDamageTargetBJ ()`
calls to the function `UnitDamageTarget ()`, and always specifies a melee
attack. When rewriting the code for both the Hold and Ancient Slates, I simply
kept this behavior, wanting to adhere to the existing code.

To be honest, I have no clue if the custom version of Spiked Carapace is
functioning properly. I believe it is not. But it doesn't matter at this
point. It is setup to not return any damage, but here we are. I can confirm,
however, that the base damage for the only affected units (Hold, Ancient, and
Damage Slates) is setup to not be considered melee.

Anyways, the triggered damage was considered melee (and now it is not), and
was having damage returned due to a bugged (and doubtful it will ever be
fixed) Spiked Carapace. The end.

Oh, I went ahead and rewrote the code for the Damage Slate, making use of Unit
Disarm. So there's that.