Fix bug that can disable an active spawn
Issue: #225

A player leaving after they have lost can disable another player's spawn. The
bug resides within the `Gem_Spawn__Stop ()` function itself and involves
multiple calls to the function. Basically, the spawn index registered to a
player is not being deallocated. In most cases, this does not present a
problem. But when a player leaks and leaves, an attempt is made to destroy the
same spawn index. But by that point, it is possible that the index may have
been recycled. Thus, an active spawn could be disabled.