Rework Poison Slate
Issue: #230

Even before the previous rework (#226), the behavior of Poison Slate had been
spotty at best. This was due to the use of `PauseUnit ()`. The rework simply
adhered to the existing behavior for the most part and shifted this pause to
use the disarm system. The previous behavior of units sometimes not being
poisoned when they should persisted, due to the slate being disarmed too often.

Both of these approaches are too much. Better is simply to check if the slate
has the required mana to cast, and, if so, check the return status of said
cast. This, in turn, sticks to the behavior the player expects: a slate that
does no damage beyond what the poison damage applies, but poisons if the slate
has the mana necessary.