Remove kill bonus spell immunity
Issue: #228, #238

The use of 'Spell Immunity' as a base for the third tier of kill bonus
abilities is the culprit behind a couple bugs. First, the inability at times
to swap structures using a Black Opal or Red Quartz Crystal has been resolved.
Second, disarm will work properly even on structures that have seventy or more
kills.

This requires a change to the base version of the map due to the need to
delete the abilities involved.