Prevent Gold Mine button shifting upon activation
Issues: #248

The goal here is to prevent the buttons from shifting. This happens when the
'Cancel' button appears, and pushes everything to the left. It can even push
something off the button row onto the middle one.

The solution is to put permanent luck changing buttons onto the middle row,
thus leaving a free button slot on the bottom row for 'Cancel' to occupy when
it appears. Of course, the remaining buttons on the bottom row must be shifted
left.