Make use of the Upkeep area
Issues: #89

This was an unutilized area that sometimes displays useful information in
other maps. It is entirely customizable, except for the percentage that
referes to the 'Current Gold Income' level.

Basic information has been put here: the map name and version; the map
maintainer; and the websites supporting the map.