bin/map: Add `debug` keyword support
Issues: #264

The `debug` keyword is specified in the JASS language syntax. It seems
to be ignored entirely by Warcraft 3, perhaps due to an internal flag
not being switched on. This commit merely extends upon this behavior
without altering the standard JASS syntax, making the `debug` keyword
actually usable.

When debugging is enabled, all `debug` keywords are removed, thus
leaving the statements they modify in the script to be executed.
However, when debugging is disabled the `debug` keywords and any
modified statements are removed entirely. This ensures that any
potential side effects (e.g. code bloat or performance concerns) are
alleviated entirely.