.github: Update the issue template
Clarify that a replay can be either a link or directly attached.
Previously, it seemed as though a ZIP file was required.

Additionally, request that the patch version used to play the game is
provided. It is entirely possible that someone is playing using an
alternative patch version, perhaps in single player or on a private
server.