share/maps: Update base map file
The import path for 'LoadingScreen.mdx' was incorrect, and has been
fixed. Also, the version for this base map has been removed. It is
simply blank in the two locations it is still set manually: the map
description and the loading screen. This change is also reflected in the
file name.