gem: Support Map 0.2.0
This makes the required changes to support Map 0.2.0. For the most part,
file lists simply needed updating to include the directory before each
file.

However, as `imports` was reworked further, it was necessary to include
two new files. One imports an entire directory, similar to the old
behavior. The other imports the 'war3map.j'. This rework allows removal
of the link that was used previously.