share/environment/header-loading-screen: Add plugin
Previously, the map description within the header and the loading screen
text was stored directly within the map. Changes required use of the
World Editor. Now it is possible to use the configuration file (see the
`map.description` and `map.loading.text` settings). This environment
plugin allows use of JASS globals to customize this text to a certain
extent.