src/unit/indexer: Ensure index `8191` is not used
Even though the stack would only use indices of the range `[0, 8190]`,
indices of the range `[1, 8191]` were used for units. As such, this had
the potential to break save/load functionality. From this point onward
unit indices will be of the range `[1, 8190]`.

Previous tests for an unrelated save/load issue did not encounter
issues. As such, I'm not entirely sure this was an issue.