src/gem/selection: Add marker module
Issues: #61

This extends upon the existing behavior for marking a selection
location. First, it moves the dummy unit if its target moves. There is a
slight delay, but this is not a big deal. Second, it removes the dummy
unit if the target dies or is removed.

The above changes to the behavior surrounding these dummy display units
resolves the issue regarding the units remaining if a player leaves.