src/selection: Slate stack checking ends prematurely
The problems is a result of the gem selection rewrite. A shift was done
from using a function call instead of creating a new trigger. As a
result, there was no trigger to execute again in the event that another
check was necessary.

This resolves the issue by returning the old invocation behavior. This
whole piece of code will need rewriting later anyway.