share/objects: Lower Anti-Stuck cooldown
Issues: #240

There is no sense in having the Anti-Stuck ability have such a long
cooldown (i.e. five minutes). Using it gains no advantage for a player;
in fact, the result is quite the opposite. True it is rare for monsters
to get stuck on each other, but it would be a shame to penalize a player
if the event were to happen twice in quick succession.