share/objects: Move Downgrade button
Issues: #168

This shift is needed, as the long term goal is to give each button a
permanent spot on the command card. The alternative solution was to
shift the initialization order in Gem Selection. This would have caused
Combine to be placed before Downgrade, and thus Downgrade would only be
pushed to the third column is Combine was possible.