share/environment: Streamline pre-release versioning
This commit attempts to simplify the pre-release versioning process.

First, whatever value is set in `Gem_Version__PRE_RELEASE` is used in
the pre-release part of the version string. The assumption is that
`alpha` implies feature incomplete, and `beta` implies feature complete.
No longer will commits be tagged as pre-release versions. To be fair,
anything could be used here. But that will be the convention used from
now on.

Second, a new global is created within the build environment only:
`Gem_Version___STRING`. This global will only be available in the
environment, and is intended for use by Grimex scripts, to simplify use
of the map's version string.