src/settings: Shift from using "\n" to " "
An empty string (i.e. `""`) is not enough to cause the functions
displaying text to print a new line. Blizzard uses a single space to
achieve this result in within the `blizzard.j`. We've been using a
single new line character (i.e. `"\n"`), which only works because it
seems to ignored. If it wasn't, we would be seeing two new lines. Let's
switch to the single space just to be safe.