share/objects: Change base unit for Topazes
Issues: #296

The previous base unit for Topazes was Naisha (`ensh`). It was found
that the use of this base unit caused issues in the map, in particular
crashes when giving consecutive orders back to back. This behavior
expressed itself when selecting and combining quickly, using a Topaz as
the focal point.

The fix for this bug is switching the base unit to be Guard Tower
(`hgtw`), whcih is what the majority of the other gem types use. Why
this fixes the problem is unknown.