share/objects: Change Elder debuff
The primary strategy in Gem (for lack of alternative options) is
stacking armor reduction. I see no problem with this so long as the
average player has access to all options.

However, once you add in kill bonuses certain structures become far too
effective. The Elder Slate is a prime example of this. It put players
who did not have the luck of getting one early enough for it to level at
an extreme disadvantage.