src/gem-3.1: Remove ownership change during placement/selection
This change was needed due to the original design of placement and
selection. The old style simply enumerated the units in the player's
control. Thus, the need to give ownership to someone else temporarily.

The rework to the placement selection system in 1.4.0 removed
the need for this behavior. One could say I forgot to remove it. Someone
else might say that it wasn't time yet.