share/maps: Update base map to have four computers
This change is necessary as it enables additional characters to be used
via HCL. With only the single computer, there was a maximum of two
guaranteed characters. With the four computeres, there is now a maximum
of five guaranteed characters. This increases the maximum number of
combinaions (when using the full range of the HCL character set) from
`1681` to `115856201`. This is probably far greater than is necessary.